+++ src/refactoring/components/admin/index.ts
@@ -1,1 +1,3 @@
 export * from "./AdminPage";
+export * from "./NewProductCard";
+export * from "./ProductCard";
